Add dict_path/index.js
